add column version (#21)

+++ src/internal/command/Command.php
@@ -7,6 +7,7 @@ namespace kuaukutsu\poc\migration\internal\command;
 use Override;
 use Throwable;
 use kuaukutsu\poc\migration\connection\ConnectionInterface;
+use kuaukutsu\poc\migration\Context;
 
 /**
  * @psalm-internal kuaukutsu\poc\migration
@@ -20,29 +21,41 @@ final readonly class Command implements CommandInterface
     }
 
     #[Override]
-    public function fetchSavedMigrationNames(Args $args = new Args()): array
+    public function fetchApplied(Args $args = new Args()): array
     {
-        // SQLSTATE[42P01]: Undefined table: 7 ERROR:  relation "migration" does not exist
-        $query = sprintf('SELECT name FROM %s ORDER BY atime DESC, name DESC', $this->params->table);
+        $params = [];
+
+        $query = sprintf('SELECT name, version FROM %s', $this->params->table);
+        if ($args->version > 0) {
+            $query .= ' WHERE version=:version';
+            $params['version'] = $args->version;
+        }
+
+        $query .= ' ORDER BY atime DESC, name DESC';
         if ($args->limit > 0) {
             $query .= ' LIMIT ' . $args->limit;
         }
 
-        return $this->connection->query($query);
+        return $this->connection->fetchRecord($query, $params);
     }
 
     #[Override]
-    public function up(string $queryString, string $filename): bool
+    public function up(Context $context): bool
     {
+        if ($context->dryRun) {
+            return false;
+        }
+
         $transaction = $this->connection->beginTransaction();
 
         try {
-            $transaction->exec($queryString);
+            $transaction->exec($context->query);
             $transaction->exec(
                 sprintf(
-                    'INSERT INTO %s (name, atime) VALUES (\'%s\', \'%s\')',
+                    'INSERT INTO %s (name, version, atime) VALUES (\'%s\', %d, \'%s\')',
                     $this->params->table,
-                    $filename,
+                    $context->filename,
+                    $context->version,
                     gmdate('Y-m-d H:i:s'),
                 )
             );
@@ -56,17 +69,21 @@ final readonly class Command implements CommandInterface
     }
 
     #[Override]
-    public function down(string $queryString, string $filename): bool
+    public function down(Context $context): bool
     {
+        if ($context->dryRun) {
+            return false;
+        }
+
         $transaction = $this->connection->beginTransaction();
 
         try {
-            $transaction->exec($queryString);
+            $transaction->exec($context->query);
             $transaction->exec(
                 sprintf(
                     'DELETE FROM %s WHERE name=\'%s\'',
                     $this->params->table,
-                    $filename,
+                    $context->filename,
                 )
             );
         } catch (Throwable $exception) {
@@ -79,12 +96,16 @@ final readonly class Command implements CommandInterface
     }
 
     #[Override]
-    public function exec(string $queryString, string $filename): bool
+    public function exec(Context $context): bool
     {
+        if ($context->dryRun) {
+            return false;
+        }
+
         $transaction = $this->connection->beginTransaction();
 
         try {
-            $transaction->exec($queryString);
+            $transaction->exec($context->query);
         } catch (Throwable $exception) {
             $transaction->rollBack();
